+++ web_src/src/pages/factories/WorkOrderDetailLoadedView.tsx
@@ -7,14 +7,13 @@ import type {
   FactoriesWorkOrderState,
 } from "@/api-client";
 import { workOrdersPath } from "./lib/factoryPagePaths";
-import { getWorkOrderDisplayKey } from "./lib/workOrderProgress";
+import { getWorkOrderDisplayKey, type WorkOrderDisplayStatus } from "./lib/workOrderProgress";
 import { factoryContentBodyClassName } from "./pages/factoryPageLayoutStyles";
 import { WorkOrderActivityTimeline } from "./WorkOrderActivityTimeline";
 import { WorkOrderCommentComposer } from "./WorkOrderCommentComposer";
 import { WorkOrderDescription } from "./WorkOrderDescription";
 import { WorkOrderDetailHeader } from "./WorkOrderDetailHeader";
 import { WorkOrderDetailSidebar } from "./WorkOrderDetailSidebar";
-import type { WorkOrderDisplayStatus } from "./lib/workOrderProgress";
 
 interface WorkOrderDetailLoadedViewProps {
   organizationId: string;
@@ -58,7 +57,33 @@ interface WorkOrderDetailLoadedViewProps {
   onAddComment: (body: string) => Promise<void>;
 }
 
-export function WorkOrderDetailLoadedView({
+export function WorkOrderDetailLoadedView(props: WorkOrderDetailLoadedViewProps) {
+  const identifier = getWorkOrderDisplayKey(props.order, props.factoryKey);
+  return (
+    <>
+      <WorkOrderDetailHeader
+        orderTitle={props.order.title ?? "Work Order"}
+        orderIdentifier={identifier === "—" ? undefined : identifier}
+        backHref={workOrdersPath(props.organizationId, props.factoryKey)}
+        displayStatus={props.displayStatus}
+        isOpen={props.isOpen}
+        isDispatchable={props.isDispatchable}
+        isClosed={props.isClosed}
+        canClose={props.canClose}
+        canManage={props.canManage}
+        isCompleting={props.isCompleting}
+        isRejecting={props.isRejecting}
+        isClosing={props.isClosing}
+        isUpdatingStatus={props.isUpdatingStatus}
+        onClose={props.onClose}
+        onStatusChange={props.onStatusChange}
+      />
+      <WorkOrderDetailBody {...props} />
+    </>
+  );
+}
+
+function WorkOrderDetailBody({
   organizationId,
   factoryKey,
   order,
@@ -78,107 +103,78 @@ export function WorkOrderDetailLoadedView({
   assigneeNames,
   factoryLines,
   canEditFactoryLines,
-  isOpen,
   isDispatchable,
-  isClosed,
   canDispatch,
-  canClose,
   canAssign,
   canManage,
   permissionsLoading,
   isDispatching,
-  isCompleting,
-  isRejecting,
-  isClosing,
   isAssigneesSaving,
-  isUpdatingStatus,
   isAddingComment,
   onDispatch,
-  onClose,
   onAssigneesSave,
-  onStatusChange,
   onAddComment,
 }: WorkOrderDetailLoadedViewProps) {
-  const identifier = getWorkOrderDisplayKey(order, factoryKey);
   return (
-    <>
-      <WorkOrderDetailHeader
-        orderTitle={order.title ?? "Work Order"}
-        orderIdentifier={identifier === "—" ? undefined : identifier}
-        backHref={workOrdersPath(organizationId, factoryKey)}
-        displayStatus={displayStatus}
-        isOpen={isOpen}
-        isDispatchable={isDispatchable}
-        isClosed={isClosed}
-        canClose={canClose}
-        canManage={canManage}
-        isCompleting={isCompleting}
-        isRejecting={isRejecting}
-        isClosing={isClosing}
-        isUpdatingStatus={isUpdatingStatus}
-        onClose={onClose}
-        onStatusChange={onStatusChange}
-      />
-      <div className={factoryContentBodyClassName}>
-        <div className="grid gap-x-[var(--workspace-column-gap)] gap-y-0 lg:grid-cols-[minmax(0,1fr)_var(--workspace-detail-sidebar-width)]">
-          <div className="min-w-0">
-            {order.description ? <WorkOrderDescription description={order.description} /> : null}
+    <div className={factoryContentBodyClassName}>
+      <div className="grid gap-x-[var(--workspace-column-gap)] gap-y-0 lg:grid-cols-[minmax(0,1fr)_var(--workspace-detail-sidebar-width)]">
+        <div className="min-w-0">
+          {order.description ? <WorkOrderDescription description={order.description} /> : null}
 
-            <section className={order.description ? "mt-10" : undefined}>
-              <h2 className="workspace-section-title">Activity</h2>
-              <p className="workspace-body-text mt-1 text-muted-foreground">
-                Actions and comments on the work order, plus factory line runs.
-              </p>
-              <div className="mt-4">
-                <WorkOrderActivityTimeline
-                  organizationId={organizationId}
-                  factoryKey={factoryKey}
-                  order={order}
-                  events={events}
-                  eventsError={eventsError}
-                  isLoading={isEventsLoading}
-                  hasMoreEvents={hasMoreEvents}
-                  isLoadingMoreEvents={isLoadingMoreEvents}
-                  onLoadMoreEvents={onLoadMoreEvents}
-                  onRetryEvents={onRetryEvents}
-                  footer={
-                    <WorkOrderCommentComposer
-                      canComment={canManage}
-                      isSubmitting={isAddingComment}
-                      onSubmit={onAddComment}
-                    />
-                  }
-                />
-              </div>
-            </section>
-          </div>
-
-          <aside className="mt-1 lg:sticky lg:top-16 lg:self-start">
-            <WorkOrderDetailSidebar
-              organizationId={organizationId}
-              factoryKey={factoryKey}
-              order={order}
-              artifacts={artifacts}
-              isArtifactsLoading={isArtifactsLoading}
-              artifactsError={artifactsError}
-              displayStatus={displayStatus}
-              statusMeta={statusMeta}
-              assigneeIds={assigneeIds}
-              assigneeNames={assigneeNames}
-              factoryLines={factoryLines}
-              canEditFactoryLines={canEditFactoryLines}
-              canAssign={canAssign}
-              canDispatch={canDispatch}
-              permissionsLoading={permissionsLoading}
-              isAssigneesSaving={isAssigneesSaving}
-              isDispatchable={isDispatchable}
-              isDispatching={isDispatching}
-              onAssigneesSave={onAssigneesSave}
-              onDispatch={onDispatch}
-            />
-          </aside>
+          <section className={order.description ? "mt-10" : undefined}>
+            <h2 className="workspace-section-title">Activity</h2>
+            <p className="workspace-body-text mt-1 text-muted-foreground">
+              Actions and comments on the work order, plus factory line runs.
+            </p>
+            <div className="mt-4">
+              <WorkOrderActivityTimeline
+                organizationId={organizationId}
+                factoryKey={factoryKey}
+                order={order}
+                events={events}
+                eventsError={eventsError}
+                isLoading={isEventsLoading}
+                hasMoreEvents={hasMoreEvents}
+                isLoadingMoreEvents={isLoadingMoreEvents}
+                onLoadMoreEvents={onLoadMoreEvents}
+                onRetryEvents={onRetryEvents}
+                footer={
+                  <WorkOrderCommentComposer
+                    canComment={canManage}
+                    isSubmitting={isAddingComment}
+                    onSubmit={onAddComment}
+                  />
+                }
+              />
+            </div>
+          </section>
         </div>
+
+        <aside className="mt-1 lg:sticky lg:top-16 lg:self-start">
+          <WorkOrderDetailSidebar
+            organizationId={organizationId}
+            factoryKey={factoryKey}
+            order={order}
+            artifacts={artifacts}
+            isArtifactsLoading={isArtifactsLoading}
+            artifactsError={artifactsError}
+            displayStatus={displayStatus}
+            statusMeta={statusMeta}
+            assigneeIds={assigneeIds}
+            assigneeNames={assigneeNames}
+            factoryLines={factoryLines}
+            canEditFactoryLines={canEditFactoryLines}
+            canAssign={canAssign}
+            canDispatch={canDispatch}
+            permissionsLoading={permissionsLoading}
+            isAssigneesSaving={isAssigneesSaving}
+            isDispatchable={isDispatchable}
+            isDispatching={isDispatching}
+            onAssigneesSave={onAssigneesSave}
+            onDispatch={onDispatch}
+          />
+        </aside>
       </div>
-    </>
+    </div>
   );
 }
